move type *Example from Simulatins module to Global
