+++ src/simulations/ABMSimulation.jl
@@ -5,6 +5,7 @@ Definition of an ABM-Simulation type.
 export ABMSimulation 
 using MultiABMs: AbstractABM
 using MultiABMs: errorstep 
+using Global:    AbstractExample, DummyExample 
 
 mutable struct ABMSimulation <: AbstractABMSimulation  
     model::AbstractABM
